Simple way to obtain the mandate store hash
The mandate store hash was printed to stdout each time a mandate store is loaded. Preparing a new mandate store file therefore required checking stdout to obtain the hash, which is a terrible approach for a GUI app. This commit replaces that with a modifier-click on the version number (Mac-only so far).

(resolves issue #126)