Future-proof version check: HTTPS  URLs
There is a known issue with HttpURLConnection that results in it not honouring HTTP redirects to HTTPS addresses. Since we intend to serve the current version directly from Github, which is HTTPS-only by now, the version check URL needs to be changed accordingly.

Note that the HTTP check URI theoretically needs to be maintained indefinitely because old versions won't get update notifications once it's redirected to HTTPS.