Support for UTF-8 mandate stores
Using UTF-8 mandate stores allows LibreOffice to autodetect the encoding by the presence of a BOM. This avoids issues with unexpected replacement characters.

I'm not 100% sure though that UTF-8 really is the way to go. After all, the debit files are always in windows-1252 encoding. However, in practice, those are apparently only very rarely edited directly -- as opposed to mandate store files, for which this happens all the time.