Add exception reporting by email
Experience has shown that users are more reluctant to call for help than previously expected. This has lead to considerable amounts of wasted/duplicated effort. Trying to report any exception right away may help reduce this problem in the future.

This code does not send emails itself. Rather, it reports the issue to the club's server by HTTP. The server then sends the email. This avoids issues with firewalls etc.