Add hidden 'advanced options' menu
Future expansion may require testing in deployed code. A hidden menu that can be enabled independent of the OS or other aspects is an easy way to do this.

For starters, the existing Mac-only hidden "calculate hash" option is moved to the new menu and thus becomes available in Win.